Refactor system settings and propeller configuration to incorporate unit preferences for EEXI limited power and shaft power. Update database model for EEXI limited power to IntegerField, enhance unit conversion methods, and ensure consistent display of units in the UI components.

+++ src/db/models/system_settings.py
@@ -10,7 +10,7 @@ class SystemSettings(BaseModel):
 
     sha_po_li = BooleanField(verbose_name="是否开启ShaPoLi功能", default=False)
 
-    eexi_limited_power = FloatField(null=True, verbose_name="EEXI 限制最大功率")
+    eexi_limited_power = IntegerField(null=True, verbose_name="EEXI 限制最大功率")
 
     class Meta:
         table_name = 'system_settings'
